pyproject.toml
[project]
name = "project-template"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "sqlalchemy>=2.0",
  "pymysql>=1.1",
  "PyJWT>=2.8",
  "httpx>=0.26",
  "mangum>=0.17",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "httpx>=0.26",
  "ruff>=0.5",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".build", "frontend/node_modules", "frontend/dist"]

[tool.ruff.lint]
select = ["E", "F", "I"]
